test: split the entry_point tests to each file to ensure minimal sys.path entries
