+++ examples/bzlmod/entry_points/pylint_deps_test.py
@@ -27,43 +27,14 @@ class ExampleTest(unittest.TestCase):
 
         super().__init__(*args, **kwargs)
 
-    def test_yamllint_entry_point(self):
-        rlocation_path = os.environ.get("YAMLLINT_ENTRY_POINT")
-        assert (
-            rlocation_path is not None
-        ), "expected 'YAMLLINT_ENTRY_POINT' env variable to be set to rlocation of the tool"
-
-        entry_point = pathlib.Path(runfiles.Create().Rlocation(rlocation_path))
-        self.assertTrue(entry_point.exists(), f"'{entry_point}' does not exist")
-
-        proc = subprocess.run(
-            [str(entry_point), "--version"],
-            check=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        self.assertEqual(proc.stdout.decode("utf-8").strip(), "yamllint 1.28.0")
-
-    def test_pylint_entry_point(self):
-        rlocation_path = os.environ.get("PYLINT_ENTRY_POINT")
-        assert (
-            rlocation_path is not None
-        ), "expected 'PYLINT_ENTRY_POINT' env variable to be set to rlocation of the tool"
-
-        entry_point = pathlib.Path(runfiles.Create().Rlocation(rlocation_path))
-        self.assertTrue(entry_point.exists(), f"'{entry_point}' does not exist")
-
-        proc = subprocess.run(
-            [str(entry_point), "--version"],
-            check=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        self.assertRegex(proc.stdout.decode("utf-8").strip(), "^pylint 2\.15\.9")
-
     def test_pylint_entry_point_deps(self):
-        rlocation_path = os.environ.get("PYLINT_ENTRY_POINT")
+        rlocation_path = os.environ.get("ENTRY_POINT")
+        assert (
+            rlocation_path is not None
+        ), "expected 'ENTRY_POINT' env variable to be set to rlocation of the tool"
+
         entry_point = pathlib.Path(runfiles.Create().Rlocation(rlocation_path))
+        self.assertTrue(entry_point.exists(), f"'{entry_point}' does not exist")
 
         with tempfile.TemporaryDirectory() as tmpdir:
             tmpdir = pathlib.Path(tmpdir)
